+++ src/main/java/cua/integra/cajasuamc/repository/AlumnoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AlumnoRepository extends JpaRepository<Alumno, Integer> {
+public interface AlumnoRepository extends JpaRepository<Alumno, Long> {
 
-    Alumno findByMatricula(int matricula);
+    Alumno findByMatricula(long matricula);
 }
